requirements.txt
fastapi
uvicorn[standard]
python-dotenv
fastapi-sqlalchemy
sqlalchemy
psycopg2-binary
alembic